fix the bug of [u16] and [32] to isize

+++ src/code_point.rs
@@ -33,13 +33,14 @@ impl CodePoints for str {
     }
 }
 
-impl<T: Copy + Debug + Into<isize> + 'static> CodePoints for [T] {
+/// T convert to `u32` and as `isize`, due to `isize` not implementation the trait `From<u16>`
+impl<T: Copy + Debug + Into<u32> + 'static> CodePoints for [T] {
     #[inline]
     fn code_point_at(&self, pos: isize) -> Option<isize> {
         if pos < 0 || pos >= self.len() as isize {
             return None;
         }
-        Some(self[pos as usize].into())
+        Some(self[pos as usize].into() as isize)
     }
 
     #[inline]
